+++ SharpSource/SharpSource.Test/SwitchDoesNotHandleAllEnumOptionsTests.cs
@@ -1006,6 +1006,30 @@ namespace ConsoleApplication1
     }
 
     [TestMethod]
+    public async Task SwitchDoesNotHandleAllEnumOptions_SwitchExpression_MissingArm()
+    {
+        var original = @"
+namespace ConsoleApplication1
+{
+    enum MyEnum
+    {
+        Fizz, Buzz, FizzBuzz
+    }
+
+    class MyClass
+    {
+        int Method(MyEnum e) => {|#0:e|} switch
+        {
+            MyEnum.Fizz => 1,
+            MyEnum.Buzz => 2
+        };
+    }
+}";
+
+        await VerifyCS.VerifyDiagnosticWithoutFix(original, VerifyCS.Diagnostic(SwitchDoesNotHandleAllEnumOptionsAnalyzer.Rule).WithMessage("Missing enum member in switched cases."));
+    }
+
+    [TestMethod]
     public async Task SwitchDoesNotHandleAllEnumOptions_NoCaseStatements_NormalUsing()
     {
         var original = @"